Replace ProductThumbnail component with ProductDefaultImage
