+++ src/components/products/modals/ProductPromotionModal.tsx
@@ -24,7 +24,7 @@ import {
 } from "@chakra-ui/react"
 import {FC, useEffect, useState} from "react"
 import {IProduct} from "types/ordercloud/IProduct"
-import ProductDefaultImage from "../list/ProductDefaultImage"
+import ProductDefaultImage from "../../shared/ProductDefaultImage"
 
 interface IProductPromotionModal {
   products?: IProduct[]
